backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
reportlab==4.1.0
qrcode==7.4.2